+++ packages/sdk-utils/package.json
@@ -2,21 +2,28 @@
   "name": "@percy/sdk-utils",
   "version": "1.0.0-beta.42",
   "license": "MIT",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "browser": "dist/bundle.js",
   "files": [
-    "index.js",
+    "dist",
     "test/helpers.js"
   ],
   "engines": {
     "node": ">=12"
   },
   "scripts": {
+    "build": "node ../../scripts/build",
     "lint": "eslint --ignore-path ../../.gitignore .",
     "test": "node ../../scripts/test",
     "test:coverage": "yarn test --coverage"
   },
   "publishConfig": {
     "access": "public"
+  },
+  "rollup": {
+    "output": {
+      "name": "PercySDKUtils"
+    }
   },
   "dependencies": {
     "@percy/logger": "^1.0.0-beta.42"
